scripts/linkmonitor/requirements.txt
beautifulsoup4==4.6.0
PyYAML==3.12
